--- web/connect/result/error.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>AutoKitteh - {{.name}} - Connection Error</title>
-    <link
-      rel="apple-touch-icon"
-      sizes="180x180"
-      href="/static/apple-touch-icon.png"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="32x32"
-      href="/static/favicon-32x32.png"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="16x16"
-      href="/static/favicon-16x16.png"
-    />
-    <link rel="manifest" href="/static/site.webmanifest" />
-    <link
-      rel="stylesheet"
-      type="text/css"
-      href="/{{.integ}}/connect/styles.css"
-    />
-    <meta name="robots" content="noindex, nofollow" />
-  </head>
-
-  <body>
-    <a href="/" target="_self">
-      <img class="banner" src="/static/banner.svg" alt="AutoKitteh" />
-    </a>
-
-    <table class="title">
-      <td><img class="logo" src="{{.logo}}" alt="Logo" /></td>
-      <td><h1>{{.name}} - Connection Error</h1></td>
-    </table>
-
-    <p class="error">{{.error}}</p>
-
-    <p class="info">
-      <button
-        class="submit-btn"
-        onclick="window.location.href = '/{{.integ}}/connect/?cid={{.cid}}&origin={{.origin}}';"
-      >
-        Try Again
-      </button>
-    </p>
-  </body>
-</html>
